pyproject.toml
[project]
name = "across-the-board"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "anthropic>=0.64.0",
    "pydantic-settings>=2.10.1",
    "streamlit>=1.48.1",
    "weaviate-client>=4.16.7",
]

[dependency-groups]
dev = [
    "huggingface>=0.0.1",
    "ipykernel>=6.30.1",
    "sentence-transformers>=5.1.0",
    "transformers>=4.55.2",
    "torch==2.2.2",
    "pre-commit>=4.3.0",
    "watchdog>=6.0.0",
    "ruff>=0.12.9",
    "numpy<2"
]
